requirements.txt
Pyrogram==1.4
tgcrypto 
ethon==1.0.7
anitopy
psutil
python-dotenv
feedparser 
apscheduler
psycopg2-binary
sqlalchemy==1.3.23
redis
requests